refactor: move mmc1 to new module
